+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed SQL error when running expire command
 - Store customer email from order on booking
 - Fixed incorrect qty when checking availability
+- Fixed unknown column error on the Bookings index when using Postgres
 
 ## 1.0.0-alpha.9 - 2018-09-05
 ### Fixed